<REFACTOR> remove unused dir and rename HostRouterHandler to HostHandler
