+++ src/main/java/com/instream/tenant/domain/host/handler/HostHandler.java
@@ -1,6 +1,5 @@
-package com.instream.tenant.domain.host.router;
+package com.instream.tenant.domain.host.handler;
 
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.server.ServerRequest;
 import org.springframework.web.reactive.function.server.ServerResponse;
@@ -8,7 +7,7 @@ import reactor.core.publisher.Mono;
 import static org.springframework.web.reactive.function.server.ServerResponse.ok;
 
 @Component
-public class HostRouterHandler {
+public class HostHandler {
     public Mono<ServerResponse> hello(ServerRequest request) {
         return ok().body(Mono.just("Hello World!"), String.class);
     }
